Create SettingsFormBackgroundImageField component
